Add files for GH Pages LaTeX support & theme
For LaTeX to be supported, MathJax-related scripts can simply be added in the header
of the HTML page.

How would that be done for our README.md / GH Page?

For minimal work and minimal amount of files added, the
_layout/default.html from the selected theme's repository (need to be
the exact theme's, it seems) can simply be copied and tweaked.

Then simply add the inline & external script such as on line 18-27 of
this particular default.html file.

Solution reference:
http://csega.github.io/mypost/2017/03/28/how-to-set-up-mathjax-on-jekyll-and-github-properly.html
(However the author also took an extra step of creating another
_includes/head.html file in addition to the _layout/default.html, which
I skipped here)